UserSync: Move last seen user hook to user sync service
